pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aas-uns-bridge"
version = "0.1.0"
description = "Daemonized integration service bridging AAS content to UNS and Sparkplug B"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AAS-UNS Bridge Contributors" }
]
keywords = ["aas", "uns", "sparkplug", "mqtt", "industry4.0", "isa95"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Log Analysis",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "paho-mqtt>=2.0,<3.0",
    "basyx-python-sdk>=2.0,<3.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "PyYAML>=6.0,<7.0",
    "watchdog>=4.0,<5.0",
    "prometheus-client>=0.20,<1.0",
    "protobuf>=4.25,<6.0",
    "typer>=0.12,<1.0",
    "httpx>=0.27,<1.0",
    "structlog>=24.0,<25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
    "types-PyYAML",
    "types-protobuf",
]

[project.scripts]
aas-uns-bridge = "aas_uns_bridge.cli:app"

[project.urls]
Homepage = "https://github.com/example/aas-uns-bridge"
Repository = "https://github.com/example/aas-uns-bridge"

[tool.hatch.build.targets.wheel]
packages = ["src/aas_uns_bridge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
